feat: move contract creation to utils
