+++ yearn/events.py
@@ -5,30 +5,11 @@ from brownie.network.event import EventDict, _decode_logs
 from joblib import Parallel, delayed
 from web3.middleware.filter import block_ranges
 
-from yearn.cache import memory
 from yearn.middleware import BATCH_SIZE
+from yearn.utils import contract_creation_block
 
 logger = logging.getLogger(__name__)
 
-
-@memory.cache()
-def contract_creation_block(address) -> int:
-    """
-    Use binary search to determine the block when a contract was created.
-    NOTE Requires access to archive state. A recommended option is Turbo Geth.
-
-    TODO Add fallback to BigQuery
-    """
-    logger.info("contract creation block %s", address)
-    height = chain.height
-    lo, hi = 0, height
-    while hi - lo > 1:
-        mid = lo + (hi - lo) // 2
-        if web3.eth.getCode(address, block_identifier=mid):
-            hi = mid
-        else:
-            lo = mid
-    return hi if hi != height else None
 
 
 def decode_logs(logs) -> EventDict:
