+++ yearn/ironbank.py
@@ -5,7 +5,7 @@ from brownie import Contract
 from brownie.network.contract import InterfaceContainer
 from joblib import Parallel, delayed
 
-from yearn.events import contract_creation_block
+from yearn.utils import contract_creation_block
 from yearn.multicall2 import multicall_matrix
 from yearn.prices import magic
